+++ defaults/viewer_response.js
@@ -1,0 +1,5 @@
+async function handler(event) {
+  const response = event.response;
+  console.log('on the edge');
+  return response;
+}