+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+=======
+## [v0.1.1](https://github.com/fujiwara/cfft/compare/v0.1.0...v0.1.1) - 2024-01-18
+- Add render sub command. by @fujiwara in https://github.com/fujiwara/cfft/pull/14
+- wait for kvs ready after creation. by @fujiwara in https://github.com/fujiwara/cfft/pull/15
+- add cfft kvs list --max-items by @fujiwara in https://github.com/fujiwara/cfft/pull/16
+- fix kvs association by @fujiwara in https://github.com/fujiwara/cfft/pull/17
+- request and response can parse a HTTP text object. by @fujiwara in https://github.com/fujiwara/cfft/pull/18
+- add test --run option by @fujiwara in https://github.com/fujiwara/cfft/pull/19
+
+## [v0.0.9](https://github.com/fujiwara/cfft/compare/v0.0.8...v0.0.9) - 2024-01-01
+- Introduce template syntax by go-config by @fujiwara in https://github.com/fujiwara/cfft/pull/9
+- Add env to testCases by @fujiwara in https://github.com/fujiwara/cfft/pull/11
+- use CloudFront KeyValueStore by @fujiwara in https://github.com/fujiwara/cfft/pull/12
+
+## [v0.0.8](https://github.com/fujiwara/cfft/compare/v0.0.7...v0.0.8) - 2023-12-24
+- Publish by @fujiwara in https://github.com/fujiwara/cfft/pull/8
+
+## [v0.0.6](https://github.com/fujiwara/cfft/compare/v0.0.5...v0.0.6) - 2023-12-23
+- files in a configuration file resolves relative path. by @fujiwara in https://github.com/fujiwara/cfft/pull/5
+
 ## [v0.0.5](https://github.com/fujiwara/cfft/compare/v0.0.4...v0.0.5) - 2023-12-23
 - Introduce config file interface. by @fujiwara in https://github.com/fujiwara/cfft/pull/3